fixed_map

Module map

Contains the fixed Map implementation.

Structs

Map

A fixed map with storage specialized through the Key trait.

Enums

Entry

A view into a single entry in a map, which may either be vacant or occupied.

Traits

MapStorage

The trait defining how storage works.

OccupiedEntry

A view into an occupied entry in a Map. It is part of the Entry enum.

VacantEntry

A view into a vacant entry in a Map. It is part of the Entry enum.

Type Aliases

IntoIter

The iterator produced by Map::into_iter.

Iter

The iterator produced by Map::iter.

IterMut

The iterator produced by Map::iter.

Keys

The iterator produced by Map::keys.

Values

The iterator produced by Map::values.

ValuesMut

The iterator produced by Map::values_mut.