fixed_map

Struct Set

pub struct Set<T>
where
    T: Key,
{ /* private fields */ }

A fixed set with storage specialized through the Key trait.

Examples

use fixed_map::{Key, Set};

#[derive(Clone, Copy, Key)]
enum Part {
    One,
    Two,
}

#[derive(Clone, Copy, Key)]
enum MyKey {
    Simple,
    Composite(Part),
    String(&'static str),
    Number(u32),
    Singleton(()),
    Option(Option<Part>),
    Boolean(bool),
}

let mut set = Set::new();

set.insert(MyKey::Simple);
set.insert(MyKey::Composite(Part::One));
set.insert(MyKey::String("foo"));
set.insert(MyKey::Number(1));
set.insert(MyKey::Singleton(()));
set.insert(MyKey::Option(None));
set.insert(MyKey::Option(Some(Part::One)));
set.insert(MyKey::Boolean(true));

assert!(set.contains(MyKey::Simple));
assert!(set.contains(MyKey::Composite(Part::One)));
assert!(!set.contains(MyKey::Composite(Part::Two)));
assert!(set.contains(MyKey::String("foo")));
assert!(!set.contains(MyKey::String("bar")));
assert!(set.contains(MyKey::Number(1)));
assert!(!set.contains(MyKey::Number(2)));
assert!(set.contains(MyKey::Singleton(())));
assert!(set.contains(MyKey::Option(None)));
assert!(set.contains(MyKey::Option(Some(Part::One))));
assert!(!set.contains(MyKey::Option(Some(Part::Two))));
assert!(set.contains(MyKey::Boolean(true)));
assert!(!set.contains(MyKey::Boolean(false)));

Implementations

impl<T> Set<T>
where T: Key,

A set implementation that uses fixed storage.

Examples

use fixed_map::{Key, Set};

#[derive(Clone, Copy, Key)]
enum MyKey {
    First,
    Second,
}

let mut m = Set::new();
m.insert(MyKey::First);

assert_eq!(m.contains(MyKey::First), true);
assert_eq!(m.contains(MyKey::Second), false);

Using a composite key:

use fixed_map::{Key, Set};

#[derive(Clone, Copy, Key)]
enum Part {
    A,
    B,
}

#[derive(Clone, Copy, Key)]
enum MyKey {
    Simple,
    Composite(Part),
}

let mut m = Set::new();
m.insert(MyKey::Simple);
m.insert(MyKey::Composite(Part::A));

assert_eq!(m.contains(MyKey::Simple), true);
assert_eq!(m.contains(MyKey::Composite(Part::A)), true);
assert_eq!(m.contains(MyKey::Composite(Part::B)), false);

pub fn new() -> Set<T>

Creates an empty Set.

Examples

use fixed_map::{Key, Set};

#[derive(Clone, Copy, Key)]
enum MyKey {
    One,
    Two,
}

let set: Set<MyKey> = Set::new();

pub fn iter(&self) -> Iter<'_, T>

An iterator visiting all values in arbitrary order. The iterator element type is T.

Examples

use fixed_map::{Key, Set};

#[derive(Debug, Clone, Copy, Key, PartialEq, Eq)]
enum MyKey {
    One,
    Two,
    Three,
}

let mut set = Set::new();
set.insert(MyKey::One);
set.insert(MyKey::Two);

assert_eq!(set.iter().collect::<Vec<_>>(), vec![MyKey::One, MyKey::Two]);

pub fn contains(&self, value: T) -> bool

Returns true if the set currently contains the given value.

Examples

use fixed_map::{Key, Set};

#[derive(Clone, Copy, Key)]
enum MyKey {
    One,
    Two,
}

let mut set = Set::new();
set.insert(MyKey::One);
assert_eq!(set.contains(MyKey::One), true);
assert_eq!(set.contains(MyKey::Two), false);

pub fn insert(&mut self, value: T) -> bool

Adds a value to the set.

If the set did not have this value present, true is returned.

If the set did have this value present, false is returned.

Examples

use fixed_map::{Key, Set};

#[derive(Clone, Copy, Key)]
enum MyKey {
    One,
    Two,
}

let mut set = Set::new();
assert!(set.insert(MyKey::One));
assert!(!set.is_empty());

set.insert(MyKey::Two);
assert!(!set.insert(MyKey::Two));
assert!(set.contains(MyKey::Two));

pub fn remove(&mut self, value: T) -> bool

Removes a value from the set. Returns true if the value was present in the set.

Examples

use fixed_map::{Key, Set};

#[derive(Clone, Copy, Key)]
enum MyKey {
    One,
    Two,
}

let mut set = Set::new();
set.insert(MyKey::One);
assert_eq!(set.remove(MyKey::One), true);
assert_eq!(set.remove(MyKey::One), false);

pub fn retain<F>(&mut self, f: F)
where F: FnMut(T) -> bool,

Retains only the elements specified by the predicate.

In other words, remove all elements e for which f(e) returns false. The elements are visited in unsorted (and unspecified) order.

Examples

use fixed_map::{Key, Set};

#[derive(Clone, Copy, Key)]
enum MyKey {
    First,
    Second,
}

let mut set = Set::new();

set.insert(MyKey::First);
set.insert(MyKey::Second);

set.retain(|k| matches!(k, MyKey::First));

assert_eq!(set.len(), 1);
assert_eq!(set.contains(MyKey::First), true);
assert_eq!(set.contains(MyKey::Second), false);

Using a composite key:

use fixed_map::{Key, Set};

#[derive(Clone, Copy, Key)]
enum MyKey {
    First(bool),
    Second(bool),
}

let mut set = Set::new();

set.insert(MyKey::First(true));
set.insert(MyKey::First(false));
set.insert(MyKey::Second(true));
set.insert(MyKey::Second(false));

let mut other = set.clone();
assert_eq!(set.len(), 4);

set.retain(|k| matches!(k, MyKey::First(true) | MyKey::Second(true)));

assert_eq!(set.len(), 2);
assert_eq!(set.contains(MyKey::First(true)), true);
assert_eq!(set.contains(MyKey::First(false)), false);
assert_eq!(set.contains(MyKey::Second(true)), true);
assert_eq!(set.contains(MyKey::Second(false)), false);

other.retain(|k| matches!(k, MyKey::First(_)));

assert_eq!(other.len(), 2);
assert_eq!(other.contains(MyKey::First(true)), true);
assert_eq!(other.contains(MyKey::First(false)), true);
assert_eq!(other.contains(MyKey::Second(true)), false);
assert_eq!(other.contains(MyKey::Second(false)), false);

pub fn clear(&mut self)

Clears the set, removing all values.

Examples

use fixed_map::{Key, Set};

#[derive(Clone, Copy, Key)]
enum MyKey {
    One,
    Two,
}

let mut set = Set::new();
set.insert(MyKey::One);
set.clear();
assert!(set.is_empty());

pub fn is_empty(&self) -> bool

Returns true if the set contains no elements.

Examples

use fixed_map::{Key, Set};

#[derive(Clone, Copy, Key)]
enum MyKey {
    One,
    Two,
}

let mut set = Set::new();
assert!(set.is_empty());
set.insert(MyKey::One);
assert!(!set.is_empty());

pub fn len(&self) -> usize

Returns the number of elements in the set.

Examples

use fixed_map::{Key, Set};

#[derive(Clone, Copy, Key)]
enum MyKey {
    First,
    Second,
}

let mut set = Set::new();
assert_eq!(set.len(), 0);
set.insert(MyKey::First);
assert_eq!(set.len(), 1);

impl<T> Set<T>
where T: Key, T::SetStorage: RawStorage,

pub fn as_raw(&self) -> <T::SetStorage as RawStorage>::Value

Get the raw value of the set.

Examples

use fixed_map::{Key, Set};

#[derive(Debug, Clone, Copy, Key)]
#[key(bitset)]
enum MyKey {
    First,
    Second,
}

let mut set = Set::new();
assert!(set.as_raw() == 0);
set.insert(MyKey::First);
assert!(set.as_raw() != 0);

let set2 = Set::from_raw(set.as_raw());
assert_eq!(set, set2);

pub fn from_raw(raw: <T::SetStorage as RawStorage>::Value) -> Self

Construct the set from a raw value.

Examples

use fixed_map::{Key, Set};

#[derive(Debug, Clone, Copy, Key)]
#[key(bitset)]
enum MyKey {
    First,
    Second,
}

let mut set = Set::new();
assert!(set.as_raw() == 0);
set.insert(MyKey::First);
assert!(set.as_raw() != 0);

let set2 = Set::from_raw(set.as_raw());
assert_eq!(set, set2);